combine generate-header-footer.js and header.js into header-footer.js
