+++ about.html
@@ -71,9 +71,7 @@
     <footer></footer>
 
     <!-- ~~~~~~~~~~~~~~~~~~~~~~~~~~ Javascript ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ -->
-    <script type="module" src="scripts/common/generate-header-footer.js"></script>
-    <script src="scripts/common/header.js"></script>
-
+    <script type="module" src="scripts/common/header-footer.js"></script>
     <script type="module" src="scripts/about.js"></script>
   </body>
 </html>
