+++ contact.html
@@ -92,9 +92,7 @@
 		<footer></footer>
 
 		<!-- ~~~~~~~~~~~~~~~~~~~~~~~~~~ Javascript ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ -->
-		<script type="module" src="scripts/common/generate-header-footer.js"></script>
-		<script src="scripts/common/header.js"></script>
-
+		<script type="module" src="scripts/common/header-footer.js"></script>
 		<script src="scripts/contact-message.js"></script>
 
 		<!-- receive form data in email-->
