+++ index.html
@@ -90,9 +90,7 @@
     <footer></footer>
 
     <!-- ~~~~~~~~~~~~~~~~~~~~~~~~~~ Javascript ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ -->
-    <script type="module" src="scripts/common/generate-header-footer.js"></script>
-    <script src="scripts/common/header.js"></script>
-
+    <script type="module" src="scripts/common/header-footer.js"></script>
     <script type="module" src="scripts/home.js"></script>
   </body>
 </html>
