+++ projects/movie-invite.html
@@ -70,7 +70,6 @@
     <footer></footer>
 
     <!-- ~~~~~~~~~~~~~~~~~~~~~~~~~~ Javascript ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ -->
-		<script type="module" src="../scripts/common/generate-header-footer.js"></script>
-    <script src="../scripts/common/header.js"></script>
+		<script type="module" src="../scripts/common/header-footer.js"></script>
   </body>
 </html>
