+++ projects/supreme-court.html
@@ -104,7 +104,6 @@
     <footer></footer>
 
 		<!-- ~~~~~~~~~~~~~~~~~~~~~~~~~~ Javascript ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ -->
-		<script type="module" src="../scripts/common/generate-header-footer.js"></script>
-    <script src="../scripts/common/header.js"></script>
+		<script type="module" src="../scripts/common/header-footer.js"></script>
   </body>
 </html>
