+++ styles/common/footer.css
@@ -31,6 +31,7 @@ footer {
 
 .nav-link.footer-nav.active, .nav-link.footer-nav:hover {
     color: var(--blue);
+    font-size: 14px;
 }
 
 footer p {
